services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sympy==1.12
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.1.0
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
nltk==3.8.1
redis==5.0.1
httpx==0.25.1
python-dotenv==1.0.0
requests==2.31.0